Heartbeat Integration
Changes:
-Heartbeat pings the "MapState" with a POST containing username, which
triggers the Pulse() on the CommandPanel of that user.
-All client input will be routed through "ControllerInputRule" by
feeding it a JSON containing the username of the client under the header
"username" and the type of command being submit under the header
"command".
-JSONWrapper no longer crashes if the JSON it is trying to parse has
only one field
-PlayerRegister no longer saves the given username in a ToLower format.
Instead, it only Trims it first.

Know Bugs:
-If a player opens the browser, signs in with a name, and the Unity
server is stopped/restarted with the page still open, the page will
resume interaction with Unity. This interaction is problematic because
it means that the client will try to give Unity commands for players
that it no longer has record of.