Implemented Game State Manager
Changes:
-Added GameState, which is a custom enum. Feel free to add more states
to the list as you see fit.
-Implemented the Game State Manager! Serves as the parent to WebServer.
When its GameState variable is modified, it broadcasts an
"OnGameStateHasChanged" message down along with the value of the new
state. In this way, any control interface objects can uniformly perform
whatever tasks are associated with each game state.
-Player Disconnects now broadcast a "OnPlayerHasDisconnected" message
upward along with the username of the player that has disconnected.

Known "Bugs":
-Newly created Control Objects (now sorted as children of WebServer) no
longer try to find a character object as soon as they are created, as
they are likely created in a scene without one already existing in it.
Because of this, they do not automatically hook up to the existing
character object in the level test scene, and as such movement commands
will not be issued to any character in particular. For right now, if you
need to test that, you can manually assign the pre-placed character  to
the Control Object's "Unit" field in the editor.