Reworked Character Select Hooks
Changes:
-Added "Character" script, to be given to the physical characters
displayed during Character Select. This script holds a reference to the
prefab of the associated character to be spawned for gameplay, a
reference to the name of this character (ex. Sniper, Brawler) and a
boolean to determine if the character has already been selected, and
therefore cannot be chosen by another player.
-Added "ChooseCharacter" case to the ControllerInputRule switch. To
choose a character, feed this rule a JSON with the command
"ChooseCharacter", the username of the player and the character you wish
to choose.
-Implemented the destruction and removal of controller objects, should
their owning player disconnect during pre-game character select. When
this occurs, it also frees any character they may have had selected to
be selected by another player. Once the Game State shifts out of
character select, controllers will no longer be destroyed on disconnect.
Similarly, new players are not able to join mid-game.