Added Launcher Scene
Changes:
-Added "Launcher.unity", a scene containing only the Game State manager
and the Web Server. This scene exists only to make sure that something
like returning to the title menu does not create another instance of the
Game State manager and Web Server, as those are persistent through
scenes.
-Added a line to the Start function of the Game State manager to
immediately load in the title menu scene, so that no time is spent in
the Launcher scene.