Implemented Map Generation Framework
Changes:
-Added "TileType" enum, which lists the possible types of tiles that can
exist on a map.
-Added "Tileset" class, which features a Dictionary<TileType,
GameObject> for linking each possible tile type to a tile prefab.
-Added "Map" class, which has parameters for a Texture2D to base the
layout of the level on, for a Tileset to use during generation, and for
a string to use as the displayed name of the map.
-Updated TileGen to feature a Dictionary<Color, TileType> for use during
generation. This works by reading each pixel value of the given Map's
Layout Texture2D, then looking the returned color value from each pixel
up in its Dictionary to return a TileType. This TileType is then looked
up in the Map's Tileset dictionary to find the correct tile prefab to
spawn. In this way, maps can be made from a very small PNG, where each
pixel represents a tile.
-Example (albeit empty) prefabs exist in the Level Assets/Level
Generation folder.