Reworked Character Select (Again)
Changes:
-Added a custom enum type called "Alliance" which determines the team of
a given character. Similarly, CharacterBehavior has been given a field
to set this for the given character.
-CharacterSelectManager now functions off of an array of a class called
"CharacterSelectEntry", which has fields for the Selection Label,
Character Prefab, and Owner. The Selection Label string is the one which
is the string identifying the desired character, bundled with the
selection command JSON. Character Prefab is the character GameObject
that will be spawned in during play. Lastly, Owner is a reference to the
control object that has selected the given character.