Implemented "Coordinate" Datatype / Updated Movement
Changes:
-Implemented a "Coordinate" class, which contains an X and a Z position.
Coordinates can be added and subtracted just like a Vector2. The main
difference here is that they feature X and Z, rather than X and Y.
-Updated the PathSelection function to operate using Coordinate data
rather than the transform position of the tiles passed in to it. This
allows for tiles to be any arbitrary physical size, as all that matters
is their coordinates in the grid.
-CharacterBehavior now features a Coordinate field, to keep track of all
units on the field. Coordinates are updated every time a unit finishes
moving to a new tile.