COMMIT: ORIGINAL SIN(GLETON)
Changes:
-Made GameStateManager and Intertitle into singletons. The former was
done such that any object can influence and read the state if needed,
and the latter was done so that any object can make a message appear
using the intertitle card, which will now also not get destroyed on load
between scenes.
-GameStateManager now also broadcasts when a state changes to ALL
gameobjects. Whether they are equipped to listen using
OnGameStateHasChanged() is a different matter entirely.
-Moved QuickWeb stuff into the Resources folder
-Moved QR to UI Scripts

To Do:
-Continue cleaning up intertitles so that they flow well and don't
overlap. I'd also like to be able to dynamically assign the width of the
content between the center brackets, among other things.